Added low memory fall back to generic dunder method
This currently is untested and will need debugging. It works by checking for low memory, then saving self.data and other.data to h5 files, indexed by the overlapping segment with the most values. The originals are then deleted to free up memory, and read back in piecewise, with the method done on them at this level, before combining and outputting.